requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-filter==25.1
django-allauth==65.16.1
requests==2.33.1
cryptography==47.0.0
whitenoise==6.9.0
gunicorn==23.0.0
psycopg[binary]>=3.2.10
